// lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      rooms: {
        Row: {
          active: boolean
          chain_family: string
          chain_id: number
          color: string | null
          contract_address: string | null
          created_at: string
          creator_id: number
          game_master_action_log: Json | null
          game_master_id: number | null
          id: number
          image_url: string | null
          name: string
          participants: number
          pvp_action_log: Json | null
          room_config: Json | null
          round_time: number
          type_id: number
          updated_at: string
        }
        Insert: {
          active?: boolean
          chain_family: string
          chain_id: number
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name: string
          participants?: number
          pvp_action_log?: Json | null
          room_config?: Json | null
          round_time?: number
          type_id: number
          updated_at?: string
        }
        Update: {
          active?: boolean
          chain_family?: string
          chain_id?: number
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id?: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name?: string
          participants?: number
          pvp_action_log?: Json | null
          room_config?: Json | null
          round_time?: number
          type_id?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "rooms_creator_id_fkey"
            columns: ["creator_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_game_master_id_fkey"
            columns: ["game_master_id"]
            isOneToOne: false
            referencedRelation: "agents"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_type_id_fkey"
            columns: ["type_id"]
            isOneToOne: false
            referencedRelation: "room_types"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never